+++ android-example/src/main/java/org/rewedigital/katana/android/example/fragment/FirstFragmentModule.kt
@@ -5,7 +5,7 @@ import org.rewedigital.katana.dsl.factory
 
 const val FRAGMENT_DEPENDENCY1 = "FRAGMENT_DEPENDENCY1"
 
-val firstFragmentModule = Module {
+val FirstFragmentModule = Module {
 
     factory(name = FRAGMENT_DEPENDENCY1) { "FRAGMENT_DEPENDENCY1" }
 }