+++ android-example/src/main/java/org/rewedigital/katana/android/example/fragment/view/FragmentActivity.kt
@@ -5,8 +5,11 @@ import androidx.appcompat.app.AppCompatActivity
 import org.rewedigital.katana.Component
 import org.rewedigital.katana.KatanaTrait
 import org.rewedigital.katana.android.example.R
-import org.rewedigital.katana.android.example.fragment.fragmentActivityModule
+import org.rewedigital.katana.android.example.fragment.FragmentActivityModule
+import org.rewedigital.katana.android.example.fragment.FragmentFactoryModule
 import org.rewedigital.katana.android.modules.ActivityModule
+import org.rewedigital.katana.androidx.fragment.KatanaFragmentFactory
+import org.rewedigital.katana.inject
 
 /**
  * @see FirstFragment
@@ -17,11 +20,17 @@ class FragmentActivity : AppCompatActivity(), KatanaTrait {
     override val component: Component = Component(
         modules = listOf(
             ActivityModule(this),
-            fragmentActivityModule
+            FragmentFactoryModule,
+            FragmentActivityModule
         )
     )
 
+    private val fragmentFactory by inject<KatanaFragmentFactory>()
+
     override fun onCreate(savedInstanceState: Bundle?) {
+        // Must be set **before** super call for Fragment instantiation after orientation change
+        supportFragmentManager.fragmentFactory = fragmentFactory
+
         super.onCreate(savedInstanceState)
 
         setContentView(R.layout.activity_fragment)