+++ buildSrc/src/main/kotlin/Dependencies.kt
@@ -3,7 +3,7 @@ object Versions {
     const val androidJunit5GradlePlugin = "1.4.0.0"
     const val androidPlugin = "3.3.2"
     const val androidXCollection = "1.1.0"
-    const val androidXFragment = "1.1.0"
+    const val androidXFragment = "1.2.0"
     const val androidXLifecycleViewModel = "2.1.0"
     const val androidXLifecycleViewModelSavedState = "1.0.0-rc03"
     const val gradleVersionsPlugin = "0.27.0"